requirements.txt
# SLS项目Python依赖清单
# 安装命令: pip install -r requirements.txt

# 核心依赖
opencv-python>=4.5.0        # 图像处理和摄像头
numpy>=1.20.0              # 数值计算
pyserial>=3.5              # 串口通信

# IR8062热像仪依赖
crcmod>=1.7                # CRC校验
cmapy>=0.6.6               # 颜色映射

# 图形界面和可视化
matplotlib>=3.5.0          # 图表绘制
tkinter                    # GUI (通常随Python安装)

# 可选依赖 (用于扩展功能)
scipy>=1.7.0               # 科学计算
pandas>=1.3.0              # 数据处理
Pillow>=8.0.0              # 图像处理增强

# 开发和调试 (生产环境可选)
# pytest>=6.0.0           # 单元测试
# flake8>=3.9.0            # 代码检查